+++ sdk/nodejs/role.ts
@@ -32,6 +32,20 @@ export class Role extends pulumi.CustomResource {
         return new Role(name, <any>state, { ...opts, id: id });
     }
 
+    /** @internal */
+    public static readonly __pulumiType = 'mysql:index/role:Role';
+
+    /**
+     * Returns true if the given object is an instance of Role.  This is designed to work even
+     * when multiple copies of the Pulumi SDK have been loaded into the same process.
+     */
+    public static isInstance(obj: any): obj is Role {
+        if (obj === undefined || obj === null) {
+            return false;
+        }
+        return obj['__pulumiType'] === Role.__pulumiType;
+    }
+
     /**
      * The name of the role.
      */
@@ -54,14 +68,7 @@ export class Role extends pulumi.CustomResource {
             const args = argsOrState as RoleArgs | undefined;
             inputs["name"] = args ? args.name : undefined;
         }
-        if (!opts) {
-            opts = {}
-        }
-
-        if (!opts.version) {
-            opts.version = utilities.getVersion();
-        }
-        super("mysql:index/role:Role", name, inputs, opts);
+        super(Role.__pulumiType, name, inputs, opts);
     }
 }